requirements.txt
discord.py-2.0.0a4373+g13359372